+++ pfnet/cpfnet.pyx
@@ -17,7 +17,6 @@ cimport cconstants
 cimport cvec
 cimport cmat
 cimport cparser_raw
-cimport cgraph
 
 from scipy import misc
 import tempfile
@@ -28,17 +27,16 @@ from libc.stdlib cimport free, malloc
 
 np.import_array()
 
-# Information
-#############
-
-info = {'graphviz': bool(cgraph.HAVE_GRAPHVIZ),
-        'raw_parser': bool(cparser_raw.HAVE_RAW_PARSER),
-        'version': str(cconstants.VERSION.decode('UTF-8'))}
-
 # Constants
 ###########
 
 PI = cconstants.PI
+
+# Information
+#############
+
+def has_raw_parser():
+    return bool(cparser_raw.HAVE_RAW_PARSER)
 
 # C pointer
 ###########
@@ -132,13 +130,13 @@ cdef Matrix(cmat.Mat* m, owndata=False):
 
 class AttributeArray(np.ndarray):
 
-    def __new__(cls,data,func=None):
+    def __new__(cls, data, func=None):
         cls.func = func
         return np.asarray(data).view(cls)
 
-    def __setitem__(self,key,value):
-        self.func(value,key)
-        np.ndarray.__setitem__(self,key,value)
+    def __setitem__(self, key, value):
+        self.func(value, key)
+        np.ndarray.__setitem__(self, key, value)
 
 # Attribute int
 ###############
@@ -179,7 +177,6 @@ include "cvargen.pyx"
 include "cbat.pyx"
 include "cnet.pyx"
 include "ccont.pyx"
-include "cgraph.pyx"
 include "cfunc.pyx"
 include "cconstr.pyx"
 include "cheur.pyx"